requirements.txt
youtube-transcript-api
streamlit
pathlib
google-generativeai
python-dotenv